# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * THREE NINE BAR Inventory System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface LoginInput {
  username: string;
  password: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  admin: 'admin',
  staff: 'staff',
} as const;

export interface User {
  id: number;
  username: string;
  /** @nullable */
  name?: string | null;
  role: UserRole;
  isActive: boolean;
  createdAt?: string;
}

export interface AuthResponse {
  user: User;
}

export type UserInputRole = typeof UserInputRole[keyof typeof UserInputRole];


export const UserInputRole = {
  admin: 'admin',
  staff: 'staff',
} as const;

export interface UserInput {
  username: string;
  password: string;
  name?: string;
  role: UserInputRole;
  isActive?: boolean;
}

export type UserUpdateRole = typeof UserUpdateRole[keyof typeof UserUpdateRole];


export const UserUpdateRole = {
  admin: 'admin',
  staff: 'staff',
} as const;

export interface UserUpdate {
  username?: string;
  password?: string;
  name?: string;
  role?: UserUpdateRole;
  isActive?: boolean;
}

export interface Ingredient {
  id: number;
  name: string;
  category: string;
  unit: string;
  stockInitial: number;
  stockMinimum: number;
  currentStock: number;
  isLowStock?: boolean;
  createdAt?: string;
}

export interface IngredientInput {
  name: string;
  category: string;
  unit: string;
  stockInitial: number;
  stockMinimum: number;
}

export interface Product {
  id: number;
  name: string;
  price: number;
  isActive: boolean;
  createdAt?: string;
}

export interface ProductInput {
  name: string;
  price: number;
  isActive?: boolean;
}

export interface RecipeDetail {
  id?: number;
  ingredientId: number;
  ingredientName: string;
  unit: string;
  quantity: number;
}

export interface RecipeDetailInput {
  ingredientId: number;
  quantity: number;
}

export interface Recipe {
  id: number;
  productId: number;
  productName: string;
  details?: RecipeDetail[];
  createdAt?: string;
}

export interface RecipeInput {
  productId: number;
  details: RecipeDetailInput[];
}

export interface Sale {
  id: number;
  saleDate: string;
  productId: number;
  productName: string;
  qty: number;
  totalPrice: number;
  createdAt?: string;
}

export interface SaleInput {
  saleDate: string;
  productId: number;
  qty: number;
}

export interface Wasting {
  id: number;
  wastingDate: string;
  ingredientId: number;
  ingredientName: string;
  qty: number;
  unit: string;
  reason: string;
  /** @nullable */
  notes?: string | null;
  createdAt?: string;
}

export interface WastingInput {
  wastingDate: string;
  ingredientId: number;
  qty: number;
  reason: string;
  notes?: string;
}

export interface StockOpname {
  ingredientId: number;
  ingredientName: string;
  unit: string;
  category: string;
  stockInitial: number;
  stockIn: number;
  stockOut: number;
  stockWasting: number;
  stockFinal: number;
  stockMinimum: number;
  isLowStock: boolean;
}

export type StockMovementMovementType = typeof StockMovementMovementType[keyof typeof StockMovementMovementType];


export const StockMovementMovementType = {
  in: 'in',
  out: 'out',
  wasting: 'wasting',
} as const;

export interface StockMovement {
  id: number;
  ingredientId: number;
  ingredientName: string;
  movementType: StockMovementMovementType;
  qty: number;
  unit: string;
  referenceType: string;
  /** @nullable */
  referenceId?: number | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface StockAddInput {
  ingredientId: number;
  qty: number;
  notes?: string;
}

export interface SalesByProduct {
  productId: number;
  productName: string;
  totalQty: number;
  totalRevenue: number;
}

export interface IngredientUsage {
  ingredientId: number;
  ingredientName: string;
  unit: string;
  totalUsed: number;
}

export interface DashboardSummary {
  totalIngredients: number;
  totalProducts: number;
  totalSalesToday: number;
  totalWastingToday: number;
  totalLowStock: number;
  recentSales: Sale[];
  salesChart: SalesByProduct[];
  ingredientUsageChart: IngredientUsage[];
}

export type ListSalesParams = {
startDate?: string;
endDate?: string;
};

export type ListWastingParams = {
startDate?: string;
endDate?: string;
};

export type ListStockMovementsParams = {
ingredientId?: number;
startDate?: string;
endDate?: string;
};

export type GetSalesByProductParams = {
startDate?: string;
endDate?: string;
};

export type GetIngredientUsageParams = {
startDate?: string;
endDate?: string;
};