# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: THREE NINE BAR Inventory System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /ingredients:
    get:
      operationId: listIngredients
      tags: [ingredients]
      summary: List ingredients
      responses:
        "200":
          description: List of ingredients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ingredient"
    post:
      operationId: createIngredient
      tags: [ingredients]
      summary: Create ingredient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IngredientInput"
      responses:
        "201":
          description: Ingredient created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ingredient"
  /ingredients/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    put:
      operationId: updateIngredient
      tags: [ingredients]
      summary: Update ingredient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IngredientInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ingredient"
    delete:
      operationId: deleteIngredient
      tags: [ingredients]
      summary: Delete ingredient
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    put:
      operationId: updateProduct
      tags: [products]
      summary: Update product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete product
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /recipes:
    get:
      operationId: listRecipes
      tags: [recipes]
      summary: List recipes
      responses:
        "200":
          description: List of recipes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Recipe"
    post:
      operationId: createRecipe
      tags: [recipes]
      summary: Create recipe with ingredients
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecipeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recipe"
  /recipes/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    get:
      operationId: getRecipe
      tags: [recipes]
      summary: Get recipe details
      responses:
        "200":
          description: Recipe with ingredients
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recipe"
    put:
      operationId: updateRecipe
      tags: [recipes]
      summary: Update recipe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecipeInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recipe"
    delete:
      operationId: deleteRecipe
      tags: [recipes]
      summary: Delete recipe
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /sales:
    get:
      operationId: listSales
      tags: [sales]
      summary: List sales
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of sales
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sale"
    post:
      operationId: createSale
      tags: [sales]
      summary: Record a sale (auto-deducts stock)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaleInput"
      responses:
        "201":
          description: Sale recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"
        "400":
          description: Insufficient stock
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /sales/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    delete:
      operationId: deleteSale
      tags: [sales]
      summary: Delete sale (restores stock)
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
    LoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, username, role, isActive]
      properties:
        id:
          type: integer
        username:
          type: string
        name:
          type: ["string", "null"]
        role:
          type: string
          enum: [admin, staff]
        isActive:
          type: boolean
        createdAt:
          type: string
...
    Ingredient:
      type: object
      required: [id, name, category, unit, stockInitial, stockMinimum, currentStock]
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        unit:
          type: string
        stockInitial:
          type: number
        stockMinimum:
          type: number
        currentStock:
          type: number
        isLowStock:
          type: boolean
        createdAt:
          type: string
    IngredientInput:
      type: object
      required: [name, category, unit, stockInitial, stockMinimum]
      properties:
        name:
          type: string
        category:
          type: string
        unit:
          type: string
        stockInitial:
          type: number
        stockMinimum:
          type: number
    Product:
      type: object
      required: [id, name, price, isActive]
      properties:
        id:
          type: integer
        name:
          type: string
        price:
          type: number
        isActive:
          type: boolean
        createdAt:
          type: string
    ProductInput:
      type: object
      required: [name, price]
      properties:
        name:
          type: string
        price:
          type: number
        isActive:
          type: boolean
    RecipeDetail:
      type: object
      required: [ingredientId, ingredientName, unit, quantity]
      properties:
        id:
          type: integer
        ingredientId:
          type: integer
        ingredientName:
          type: string
        unit:
          type: string
        quantity:
          type: number
    RecipeDetailInput:
      type: object
      required: [ingredientId, quantity]
      properties:
        ingredientId:
          type: integer
        quantity:
          type: number
    Recipe:
      type: object
      required: [id, productId, productName]
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        details:
          type: array
          items:
            $ref: "#/components/schemas/RecipeDetail"
        createdAt:
          type: string
    RecipeInput:
      type: object
      required: [productId, details]
      properties:
        productId:
          type: integer
        details:
          type: array
          items:
            $ref: "#/components/schemas/RecipeDetailInput"
    Sale:
      type: object
      required: [id, saleDate, productId, productName, qty, totalPrice]
      properties:
        id:
          type: integer
        saleDate:
          type: string
        productId:
          type: integer
        productName:
          type: string
        qty:
          type: integer
        totalPrice:
          type: number
        createdAt:
          type: string
    SaleInput:
      type: object
      required: [saleDate, productId, qty]
      properties:
        saleDate:
          type: string
        productId:
          type: integer
        qty:
          type: integer
...